requirements.txt
omegaconf>=2.1.2
scikit-learn
pytorch-lightning==1.9.4
timm==0.5.4
nltk~=3.7.0
lmdb~=1.3.0
Pillow~=9.2.0
imgaug~=0.4.0
hydra-core~=1.2.0
fvcore~=0.1.5.post20220512
ray[tune]~=1.13.0
ax-platform~=0.2.5.1
PyYAML~=6.0.0
tqdm~=4.64.0
setuptools~=59.5.0
opencv-python~=4.5.5
opendatasets
transformers==4.17.0
seaborn
ftfy